// modules/mmorpg/src/model/generated/server/demo/protocol/messages.ts
export interface MapInstanceSnapshot {
  mapInstanceId: bigint;
  mapConfigId: number;
  mapHostName: string;
  dynamic: boolean;
  mapHost: MapHostEndpoint;
}
// ...
export interface MapHostEndpoint {
  name: string;
  ip: string;
  port: number;
  protocol: string;
  audience: string;
}
// ...
export interface S2L_ResolveMapInstance extends IRequest {
  rpcId?: number;
  mapInstanceId: bigint;
}
// ...
export interface L2S_ResolveMapInstance extends IResponse {
  message?: string;
  error?: number;
  rpcId?: number;
  found: boolean;
  instance: MapInstanceSnapshot;
}
// ...
export interface S2L_RegisterMapInstance extends IRequest {
  rpcId?: number;
  instance: MapInstanceSnapshot;
  ownerGeneration: bigint;
  leaseTimeoutMs: number;
}
// ...
export interface L2S_RegisterMapInstance extends IResponse {
  message?: string;
  error?: number;
  rpcId?: number;
  instance: MapInstanceSnapshot;
  created: boolean;
}
// ...
export interface S2L_RemoveMapInstance extends IRequest {
  rpcId?: number;
  mapInstanceId: bigint;
  expectedMapHostName: string;
  expectedOwnerGeneration: bigint;
}
// ...
export interface L2S_RemoveMapInstance extends IResponse {
  message?: string;
  error?: number;
  rpcId?: number;
  removed: boolean;
}